// app/types/umkm.ts
export interface Umkm {
  id: number;
  nama: string;
  nama_usaha: string;
  produk: string;
  kategori: string;
  alamat: string;
  rt: number;
  rw: number;
  status_nib: string;
  status_pelatihan: string;
  desil: number;
  status_validasi: string;
  kecamatan: string;
  tahun_laporan: number;
  latitude: number;
  longitude: number;
  url?: string;
}